LD4-35 Added security check system. Tokens are now renew automatically(send in header with response) - in case of user's role changed In case of logout - token will be deleted, a no client can pass by it. Also system can count online users precisely and adequate
SQL for DB:

create table main.access_tokens
(
	id bigint not null
		constraint access_tokens_pkey
			primary key,
	token varchar(1000) not null,
	expiration_date timestamp default CURRENT_DATE not null,
	state varchar(7) default 'ONLINE'::character varying not null,
	user_id bigint not null,
	new_token varchar
);

+++ src/main/java/dk/ledocsystem/ledoc/model/dashboard/SuperAdminStatistic.java
@@ -11,5 +11,5 @@ public class SuperAdminStatistic {
 
     private UserStat userStat;
 
-    private Integer usersOnline;
+    private Long usersOnline;
 }
